refactor(parser/boolean): consolidate boolean.go and interface.go into interface.go
